# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    LoginData:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, legal, finance, sales]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
          nullable: true
...
    ContentItem:
      type: object
      required: [id, type, title, createdAt]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [Film, TVSeries, TBN_FAST, TBN_Linear, WoF_FAST]
        title:
          type: string
        description:
          type: string
          nullable: true
        year:
          type: integer
          nullable: true
        seasons:
          type: array
          items:
            $ref: '#/components/schemas/Season'
        contractCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Season:
      type: object
      required: [id, contentItemId, seasonNumber]
      properties:
        id:
          type: string
        contentItemId:
          type: string
        seasonNumber:
          type: integer
        title:
          type: string
          nullable: true
        year:
          type: integer
          nullable: true
        episodeCount:
          type: integer
          nullable: true
    CreateContentItemRequest:
      type: object
      required: [type, title]
      properties:
        type:
          type: string
          enum: [Film, TVSeries, TBN_FAST, TBN_Linear, WoF_FAST]
        title:
          type: string
        description:
          type: string
          nullable: true
        year:
          type: integer
          nullable: true
        seasons:
          type: array
          items:
            type: object
            required: [seasonNumber]
            properties:
              seasonNumber:
                type: integer
              title:
                type: string
                nullable: true
              year:
                type: integer
                nullable: true
              episodeCount:
                type: integer
                nullable: true
    UpdateContentItemRequest:
      type: object
      properties:
        type:
          type: string
          enum: [Film, TVSeries, TBN_FAST, TBN_Linear, WoF_FAST]
        title:
          type: string
        description:
          type: string
          nullable: true
        year:
          type: integer
          nullable: true
        seasons:
          type: array
          items:
            type: object
            required: [seasonNumber]
            properties:
              seasonNumber:
                type: integer
              title:
                type: string
                nullable: true
              year:
                type: integer
                nullable: true
              episodeCount:
                type: integer
                nullable: true
...
    ContractListItem:
      type: object
      required: [id, direction, status, endType, createdAt]
      properties:
        id:
          type: string
        direction:
          type: string
          enum: [rights_in, rights_out]
        partnerId:
          type: string
        partnerName:
          type: string
        licensor:
          type: string
          nullable: true
        licensee:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, active, expired, in_perpetuity, terminated]
        startDate:
          type: string
          format: date
          nullable: true
        endType:
          type: string
          enum: [date, perpetuity, auto_renew]
        endDate:
          type: string
          format: date
          nullable: true
        territories:
          type: array
          items:
            type: string
        distributionTypes:
          type: array
          items:
            type: string
        royaltyType:
          type: string
          nullable: true
        contentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    PaginatedContent:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ContentItem'
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      operationId: login
      summary: Login with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginData'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /auth/me:
    get:
      operationId: getMe
      summary: Get current user
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /content:
    get:
      operationId: listContent
      summary: List content catalog items
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [Film, TVSeries, TBN_FAST, TBN_Linear, WoF_FAST]
      responses:
        '200':
          description: Paginated content list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedContent'
    post:
      operationId: createContent
      summary: Create a content catalog item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContentItemRequest'
      responses:
        '201':
          description: Created content item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentItem'
  /content/{id}:
    get:
      operationId: getContent
      summary: Get content item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Content item details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentItem'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      operationId: updateContent
      summary: Update content item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContentItemRequest'
      responses:
        '200':
          description: Updated content item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentItem'
    delete:
      operationId: deleteContent
      summary: Delete content item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /content/{id}/contracts:
    get:
      operationId: getContentContracts
      summary: Get all contracts linked to a content item (reverse lookup)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Contracts linked to this content
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContractListItem'
...